chore(device-discovery): rename hp_comware driver to comware
Renames `custom_napalm/hp_comware.py` → `comware.py` and the corresponding
test directory, updating all module-level imports and docstrings. The
Netmiko device-type string and ntc-templates platform string (`hp_comware`)
are external identifiers and remain unchanged.
