+++ device-discovery/custom_napalm/__init__.py
@@ -10,7 +10,7 @@ from custom_napalm.fastiron import FastIronDriver
 from custom_napalm.ftd_ssh import FTDSSHDriver
 from custom_napalm.ftos import FTOSDriver
 from custom_napalm.gaia import GaiaDriver
-from custom_napalm.hp_comware import ComwareDriver
+from custom_napalm.comware import ComwareDriver
 from custom_napalm.hp_procurve import ProCurveDriver
 from custom_napalm.huawei_vrp import VRPDriver
 from custom_napalm.netiron import NetIronDriver
