+++ device-discovery/custom_napalm/comware.py
@@ -14,7 +14,6 @@ without templates (display version).
 import ipaddress
 import logging
 import re
-import socket
 
 import napalm.base as _napalm_base
 from napalm.base import models
