+++ device-discovery/tests/custom_drivers/comware/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for custom_napalm.comware.ComwareDriver."""
