+++ device-discovery/tests/custom_drivers/comware/mock_data/test_get_config/normal/display_current-configuration.txt
