+++ device-discovery/tests/custom_drivers/comware/mock_data/test_get_config_sanitized/normal/display_saved-configuration.txt
