+++ device-discovery/tests/custom_drivers/comware/mock_data/test_get_facts/normal/display_version.txt
