+++ device-discovery/tests/custom_drivers/comware/mock_data/test_get_interfaces/normal/display_interface.txt
