+++ device-discovery/tests/custom_drivers/comware/mock_data/test_get_interfaces_ip/normal/expected_result.json
