+++ device-discovery/tests/custom_drivers/comware/mock_data/test_get_vlans/normal/expected_result.json
