--- device-discovery/tests/custom_drivers/hp_comware/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for custom_napalm.hp_comware.ComwareDriver."""
